+++ webrtc/modules/rtp_rtcp/source/producer_fec_unittest.cc
@@ -33,7 +33,7 @@ void VerifyHeader(uint16_t seq_num,
   uint32_t parsed_timestamp = (data[4] << 24) + (data[5] << 16) +
       (data[6] << 8) + data[7];
   EXPECT_EQ(timestamp, parsed_timestamp);
-  EXPECT_EQ(fec_pltype, data[kRtpHeaderSize]);
+  EXPECT_EQ(static_cast<uint8_t>(fec_pltype), data[kRtpHeaderSize]);
 }
 
 class ProducerFecTest : public ::testing::Test {
